Hashbang DOCTYPE
This is an interesting concept. I like the idea of it being 'server-like' in the first place too. This feels kind of cheeky like early Svelte versions though (not a dig on Svelte!), where this might just be a 'cool idea' rather than one that makes sense in the long term. I'm going to try it for the meantime though, as it seems to make sense in terms of IDE and TS type checking support too. I like how simple this is. The other option was to do something like `/// <DOCTYPE html/>` like a TS directive, I don't like that as much though. Another one would be to have `<DOCTYPE html/>;` as a prefix JSX element to start the page with. That kind of words, just feels a little off.