Dynamic JSX Import
Okay now importing the JSX with `import()` still allows it to work, without needing to hardcode that file into the server itself.

The next goal is to get the hashbang out properly. Or should it just be HTML5 by default? Hmm.

Oh yeah, and to run this I am now doing `tsx ./server.ts ./index.tsx`. Previously there was no argument to the server, that's new.

I think PHP runs on every request too, so I need to add support for that.