Component Reloading
Okay this is a bit of a hack. In theory, maybe this should actually just be a function afterall, since wanting to rerun it multiple times, you'd have to rebuild the module tree over and over again. So maybe instead of doing that, it should just return a function afterall.